pdf_archiver/requirements.txt
google-genai
playwright
python-dotenv
python-slugify
pypdf
aiohttp
flask
